# openapi.yaml
openapi: 3.0.3
info:
  title: YOLO Training Visualization Platform API
  version: 1.0.0
  description: |
    REST API for YOLO Training Visualization Platform - A comprehensive tool for dataset management, 
    model training, testing, validation, export, and TCP-based image processing.
    
    ## Modules
    - **IDataset**: Dataset management and Label Studio integration
    - **ITraining**: Training task creation, execution and monitoring
    - **IModel**: Model testing, validation, export and Triton deployment
    - **IImageProcessor**: TCP image processing service integration
    
  contact:
    name: chzane
    email: slxzane@outlook.com
    url: https://github.com/chzane/YoloTrainingVisualizationPlatform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:10799
    description: Local development server
paths:
...
  /ITraining/getTask:
    get:
      summary: Read a specific task config
      parameters:
        - in: query
          name: filename
          schema: { type: string }
          required: true
      responses:
        '200':
          description: Task config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StdResponse'
  /ITraining/createTask:
    post:
      summary: Create a training task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                taskName: { type: string }
                taskDescription: { type: string }
                datasetPath: { type: string }
                trainingType: { type: integer, enum: [0,1] }
                baseModelID: { type: string, description: Required when trainingType=0 }
                modelYamlFile: { type: string, description: YAML text when trainingType=1 }
                epochs: { type: integer, default: 100 }
                batchSize: { type: integer, default: 16 }
                imgSize: { type: integer, default: 640 }
                device: { type: string, enum: [cpu, gpu, gpu_idlefirst, mps] }
                gpuCUDAIndex: { type: string }
                gpuCUDANum: { type: integer }
                trainSeed: { type: integer, default: 0 }
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StdResponse'
  /ITraining/deleteTask:
    post:
      summary: Delete a task YAML by relative path under tasks directory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path: { type: string }
              required: [path]
      responses:
        '200':
          description: Delete status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StdResponse'
  /ITraining/startTask:
    post:
      summary: Start a training task execution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                taskID: { type: string }
                filename: { type: string }
                taskname: { type: string }
              required: [taskID, filename, taskname]
      responses:
        '200':
          description: Started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StdResponse'
  /ITraining/getTaskLog:
    get:
      summary: Get terminal logs for a running training task
      parameters:
        - in: query
          name: filename
          schema: { type: string }
          required: true
      responses:
        '200':
          description: Logs and running flag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StdResponse'
...
  /ITraining/getTrainingTasksHistory:
    get:
      summary: List history entries for a task
      parameters:
        - in: query
          name: taskID
          schema: { type: string }
          required: true
      responses:
        '200':
          description: History
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StdResponse'
  /ITraining/getTrainingTaskOutputFile:
    get:
      summary: Read a training output file content or image
      parameters:
        - in: query
          name: taskID
          schema: { type: string }
          required: true
        - in: query
          name: filePath
          schema: { type: string }
          required: true
        - in: query
          name: resultFilePath
          schema: { type: string }
          required: true
      responses:
        '200':
          description: Output file content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StdResponse'
...
components:
  schemas:
    StdResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        msg:
          type: string
          example: OK
        data:
          description: Response payload (varies by endpoint)
          nullable: true